# app/http/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "zhihu测试",
        "title": "zhihu",
        "termsOfService": "https://github.com/swaggo/swag",
        "contact": {
            "name": "choi",
            "email": "choi"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/answer/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建回答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qa"
                ],
                "summary": "创建回答",
                "parameters": [
                    {
                        "description": "创建回答参数",
                        "name": "answerCreateParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qa.answerCreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/answer/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建回答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qa"
                ],
                "summary": "创建回答",
                "parameters": [
                    {
                        "description": "删除id",
                        "name": "answerDeleteParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qa.answerDeleteParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/question/detail": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取问题详情，包括问题的所有回答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qa"
                ],
                "summary": "获取问题详情",
                "parameters": [
                    {
                        "description": "问题id",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qa.questionDetailParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "问题详情，带回答和作者",
                        "schema": {
                            "type": "QuestionDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "qa.answerCreateParam": {
            "type": "object",
            "required": [
                "context",
                "question_id"
            ],
            "properties": {
                "context": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "qa.answerDeleteParam": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "qa.questionDetailParam": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}